Extract, test and fix RepoTermLoaderStream
